# ARROW-6351: [Ruby] Improve Arrow#values performance