--- a/ruby/red-arrow/ext/arrow/converters.cpp
+++ b/ruby/red-arrow/ext/arrow/converters.cpp
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+#include "converters.hpp"
+
+namespace red_arrow {
+  VALUE ArrayValueConverter::convert(const arrow::ListArray& array,
+                                     const int64_t i) {
+    return list_array_value_converter_->convert(array, i);
+  }
+
+  VALUE ArrayValueConverter::convert(const arrow::StructArray& array,
+                                     const int64_t i) {
+    return struct_array_value_converter_->convert(array, i);
+  }
+
+  VALUE ArrayValueConverter::convert(const arrow::UnionArray& array,
+                                     const int64_t i) {
+    return union_array_value_converter_->convert(array, i);
+  }
+
+  VALUE ArrayValueConverter::convert(const arrow::DictionaryArray& array,
+                                     const int64_t i) {
+    return dictionary_array_value_converter_->convert(array, i);
+  }
+}