@mrkn 2019-08-26
How about extracting the common parts of `red_arrow::RawValuesBuilder` and `red_arrow::RawRecordsBuilder` as a base class?  I think the both class shares a lot of common code.