@kou 2019-08-26
Extracted as `red_arrow::Converter`.