@mrkn 2019-08-26
Can `Visit` functions be moved in `red_arrow::Converter`?