@kou 2019-08-26
We can't. Because it depends on `template <typename ArrayType> void convert()`.