@mrkn 2019-08-26
What a stupid I am!